Use GearRepository in gear list component and clean up the component.
