+++ src/components/gear-list/gear-list.html
@@ -1,4 +1,4 @@
-<table id="equipment-list-container">
+<table id="gear-list-container">
     <thead>
         <tr>
             <th>Item</th>
