+++ src/components/pages/gear/gear.component.html
@@ -1,3 +1,3 @@
 <h1 class="mt-1">Gear</h1>
 <p>TODO: Add a header or buttons to select Equipment | Weapons | Industrail Equipment | Patches | etc.</p>
-<equipment-list></equipment-list>
+<gear-list></gear-list>
